large_language_model/groupchat_app_src/backend/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy>=2.0.0
asyncmy
passlib[bcrypt]
python-jose[cryptography]
python-dotenv
httpx
jinja2
